requirements.txt
PyYAML
jsonref
Faker